+++ WMS.Application/DTOs/Addresses/UserAddressResponceDto.cs
@@ -2,7 +2,7 @@ namespace WMS.Application.DTOs.UserAddresses;
 
 public class UserAddressResponseDto : UserAddressRequestDto
 {
-    public long AddressId { get; set; }
+    public int AddressId { get; set; }
     public string StateName { get; set; } = string.Empty;
     public string CityName { get; set; } = string.Empty;
