+++ WMS.Application/DTOs/Order/AdminOrderRequestDto.cs
@@ -4,7 +4,7 @@ public class AdminOrderRequestDto : CommonFilterDto{
     public bool? OnlyPending  {get;set;}
     public bool? onlypendingandaccepted  {get;set;}
     
-    public long? StateId { get; set; }
-    public long? CityId { get; set; }
+    public int? StateId { get; set; }
+    public int? CityId { get; set; }
 }
