+++ WMS.Application/DTOs/Order/OrderItemResponseDto.cs
@@ -1,7 +1,7 @@
 public class OrderItemResponseDto
 {
-    public long OrderItemId { get; set; }
-    public long ProductId { get; set; }
+    public int OrderItemId { get; set; }
+    public int ProductId { get; set; }
     public string ProductName { get; set; } = string.Empty;
     public int Quantity { get; set; }
     public decimal WeightKg { get; set; }
