+++ WMS.Domain/Entities/OrderItem.cs
@@ -4,8 +4,8 @@ namespace WMS.Domain.Entities;
 
 public class OrderItem : BaseEntity
 {
-    public long OrderId { get; set; }
-    public long ProductId { get; set; }
+    public int OrderId { get; set; }
+    public int ProductId { get; set; }
     public int Quantity { get; set; }
     public decimal WeightKg { get; set; }
     public decimal Price {get;set;}
